Encode codec names much more efficiently
By not having to add a whole bunch of padding so that every codec name is just as long as the longest one, and instead use a codec name strtab alongside offests into said table, the total size of the database has been just about halfed.
Additionally, this change now allows for a further indirection and free arrangement of relative string order which is now being used to group multiple names of the same codec together, and having the iconv_getcodecname() (now renamed to iconv_getcodecnames()) return a \0\0-terminated list of all names for a given codec, sorted in the same order by which the names appear in the input database file.