Add a bunch more stuff to the transliteration database
Also: add support for a few alias characters in the VGA terminal driver (these characters are generated by nano)

Also: Adjust the way that transliterations are generated to be more git-friendly (from now on, adding/removing transliterations won't cause the entire pre-generated database file to be modifed, but only have a couple of lines be changed). Additionally, this way it's much easier to see the patterns within the generated database!